Reduce dependency on selectedDevices
A lot of legacy code back from when Gadgetbridge only supported a single
device still depends on selectedDevices. This can lead to issues if
multiple devices of the same type are paired / connected, as it is not
possible for coordinators to know which device is being checked, and
they will simply use the first they find.